refactor(common): consolidate MessageMetadata and message types into a single file
Move MessageMetadata, ActiveSelection, UserEdits, BashOutputs, and Review types
from packages/livekit/src/types.ts and packages/common/src/base/message-context.ts
into a single consolidated packages/common/src/base/message.ts file, and re-export
them from @getpochi/common. Update all consumers to import from the new unified file.
