+++ packages/livekit/src/types.ts
@@ -2,6 +2,7 @@ import type { LanguageModelV3 } from "@ai-sdk/provider";
 import type {
   ActiveSelection,
   BashOutputs,
+  MessageMetadata,
   Review,
   UserEdits,
 } from "@getpochi/common";
@@ -9,26 +10,12 @@ import { Environment } from "@getpochi/common";
 import { GoogleVertexModel } from "@getpochi/common/configuration";
 import { type ClientTools, McpTool } from "@getpochi/tools";
 import type { Store } from "@livestore/livestore";
-import type { FinishReason, InferUITools, UIMessage } from "ai";
+import type { InferUITools, UIMessage } from "ai";
 import z from "zod";
 import type { defaultCatalog } from "./livestore";
 import type { tables } from "./livestore/default-schema";
 
-export const ZodMetadata = z.discriminatedUnion("kind", [
-  z.object({
-    kind: z.literal("assistant"),
-    totalTokens: z.number(),
-    finishReason: z.custom<FinishReason>(),
-    startedAt: z.coerce.date().optional(),
-    finishedAt: z.coerce.date().optional(),
-  }),
-  z.object({
-    kind: z.literal("user"),
-    compact: z.boolean().optional(),
-  }),
-]);
-
-export type Metadata = z.infer<typeof ZodMetadata>;
+export type Metadata = MessageMetadata;
 
 export type DataParts = {
   checkpoint: {
